+++ collector-bitmovin-player-v1/src/main/java/com/bitmovin/analytics/bitmovin/player/features/BitmovinSegmentTrackingAdapter.kt
@@ -15,7 +15,7 @@ class BitmovinSegmentTrackingAdapter(private val player: BitmovinPlayer, private
     private val observableSupport = ObservableSupport<OnDownloadFinishedEventListener>()
     private val onDownloadFinishedListener = OnDownloadFinishedListener {
         val segmentType = mapHttpRequestType(it.downloadType)
-        val segmentInfo = Segment(it.timestamp, segmentType, it.url, it.lastRedirectLocation, it.httpStatus, it.downloadTime, it.size, it.isSuccess)
+        val segmentInfo = Segment(it.timestamp, segmentType, it.url, it.lastRedirectLocation, it.httpStatus, it.downloadTime, null, it.size, it.isSuccess)
         observableSupport.notify { listener -> listener.onDownloadFinished(OnDownloadFinishedEventObject(segmentInfo)) }
     }
 
@@ -31,6 +31,10 @@ class BitmovinSegmentTrackingAdapter(private val player: BitmovinPlayer, private
             HttpRequestType.MANIFEST_HLS_MASTER -> SegmentType.MANIFEST_HLS_MASTER
             HttpRequestType.MANIFEST_HLS_VARIANT -> SegmentType.MANIFEST_HLS_VARIANT
             HttpRequestType.MANIFEST_SMOOTH -> SegmentType.MANIFEST_SMOOTH
+            HttpRequestType.KEY_HLS_AES -> SegmentType.KEY_HLS_AES
+            HttpRequestType.MEDIA_AUDIO -> SegmentType.MEDIA_AUDIO
+            HttpRequestType.MEDIA_SUBTITLES -> SegmentType.MEDIA_SUBTITLES
+            HttpRequestType.MEDIA_VIDEO -> SegmentType.MEDIA_VIDEO
             else -> SegmentType.UNKNOWN
         }
     }