+++ collector/src/main/java/com/bitmovin/analytics/LicenseKeyProvider.kt
@@ -1,0 +1,5 @@
+package com.bitmovin.analytics
+
+interface LicenseKeyProvider {
+    val licenseKey: String
+}