+++ collector/src/test/java/com/bitmovin/analytics/license/LicenseCallTests.kt
@@ -2,6 +2,7 @@ package com.bitmovin.analytics.license
 
 import android.net.Uri
 import com.bitmovin.analytics.BitmovinAnalyticsConfig
+import com.bitmovin.analytics.features.errordetails.ErrorDetailTrackingConfig
 import com.bitmovin.analytics.utils.HttpClient
 import io.mockk.every
 import io.mockk.mockk
@@ -34,37 +35,52 @@ class LicenseCallTests {
         return LicenseCall(BitmovinAnalyticsConfig(""), mockk(relaxed = true))
     }
 
-    private fun getGrantedResponseBody(settings: String) = "{\"status\": \"granted\"$settings}"
+    private fun getGrantedResponseBody(features: String) = "{\"status\": \"granted\"$features}"
 
-    private fun verifyLicenseResponse(responseBody: String, expectedSuccess: Boolean, expectedSettings: Map<String, String>?) {
+    private fun verifyLicenseResponse(responseBody: String, expectedSuccess: Boolean, expectedFeatures: FeatureConfigContainer?) {
         val licenseCall = createLicenseCall(responseBody)
         val callback = mockk<AuthenticationCallback>(relaxed = true)
         licenseCall.authenticate(callback)
-        verify { callback.authenticationCompleted(expectedSuccess, expectedSettings) }
+        verify { callback.authenticationCompleted(expectedSuccess, expectedFeatures) }
     }
 
     @Test
-    fun testLicenseResponseShouldSuccessfullyBeParsedWithoutSettings() {
+    fun testLicenseResponseShouldSuccessfullyBeParsedWithoutFeatures() {
         verifyLicenseResponse(getGrantedResponseBody(""), true, null)
     }
 
     @Test
-    fun testLicenseResponseShouldSuccessfullyBeParsedWithNullSettings() {
-        verifyLicenseResponse(getGrantedResponseBody(", \"settings\": null"), true, null)
+    fun testLicenseResponseShouldSuccessfullyBeParsedWithNullFeatures() {
+        verifyLicenseResponse(getGrantedResponseBody(", \"features\": null"), true, null)
     }
 
     @Test
-    fun testLicenseResponseShouldSuccessfullyBeParsedWithEmptySettings() {
-        verifyLicenseResponse(getGrantedResponseBody(", \"settings\": {}"), true, HashMap())
+    fun testLicenseResponseShouldSuccessfullyBeParsedWithEmptyFeatures() {
+        verifyLicenseResponse(getGrantedResponseBody(", \"features\": {}"), true, FeatureConfigContainer(null))
     }
 
     @Test
-    fun testLicenseResponseShouldSuccessfullyBeParsedWithCorrectSettings() {
-        verifyLicenseResponse(getGrantedResponseBody(", \"settings\": { \"feature\": \"foo\" }"), true, hashMapOf("feature" to "foo"))
+    fun testLicenseResponseShouldSuccessfullyBeParsedWithErrorTracking() {
+        verifyLicenseResponse(getGrantedResponseBody(", \"features\": { \"errorSegments\": {} }"), true, FeatureConfigContainer(ErrorDetailTrackingConfig(false)))
     }
 
     @Test
-    fun testLicenseResponseShouldFailWithWrongSettings() {
-        verifyLicenseResponse(getGrantedResponseBody(", \"settings\": \"asdf\""), false, null)
+    fun testLicenseResponseShouldSuccessfullyBeParsedWithDisabledErrorTracking() {
+        verifyLicenseResponse(getGrantedResponseBody(", \"features\": { \"errorSegments\": {\"enabled\": false} }"), true, FeatureConfigContainer(ErrorDetailTrackingConfig(false)))
+    }
+
+    @Test
+    fun testLicenseResponseShouldSuccessfullyBeParsedWithEnabledErrorTracking() {
+        verifyLicenseResponse(getGrantedResponseBody(", \"features\": { \"errorSegments\": {\"enabled\": true, \"numberOfSegments\": 12} }"), true, FeatureConfigContainer(ErrorDetailTrackingConfig(true, 12)))
+    }
+
+    @Test
+    fun testLicenseResponseShouldSuccessfullyBeParsedWithEnabledErrorTrackingAndTypo() {
+        verifyLicenseResponse(getGrantedResponseBody(", \"features\": { \"errorSegments\": {\"enabled\": true, \"numberOfSeeegments\": 12} }"), true, FeatureConfigContainer(ErrorDetailTrackingConfig(true)))
+    }
+
+    @Test
+    fun testLicenseResponseShouldFailWithWrongFeatures() {
+        verifyLicenseResponse(getGrantedResponseBody(", \"features\": \"asdf\""), false, null)
     }
 }